// lib/types/components/iconButton.d.ts
import { type ReactNode } from "react";
import { type TColors } from "../generic";

export type TIconButtonVariant = "filled" | "outlined" | "text" | "gradient";
export type TIconButtonSize = "sm" | "md" | "lg";
export type TIconButtonColors = TColors;
export type TIconButtonIcon = ReactNode;
export type TIconButtonClassname = string;
export type TIconButtonLoading = boolean;